Small cleanup in Tailwind
